+++ backend/routes/api/index.js
@@ -1,7 +1,7 @@
 const router = require('express').Router();
 const sessionRouter = require('./session.js');
 const usersRouter = require('./users.js');
-const businessRouter = require('./businesses');
+const businessRouter = require('./business');
 
 router.use('/session', sessionRouter);
 router.use('/users', usersRouter);
